Criando o terceiro endpoint Estatistica e removendo alguns conteudos não necessários

+++ src/main/java/br/com/thiagokenji/apiREST/model/Transacao.java
@@ -1,11 +1,9 @@
 package br.com.thiagokenji.apiREST.model;
 
-import java.sql.Date;
 import java.time.OffsetDateTime;
 
 public class Transacao {
-	//private Long id;
-	private  float valor;
+	private  double valor;
 	private OffsetDateTime  dataHora;
 	
 	
@@ -13,18 +11,11 @@ public class Transacao {
 	
 	public Transacao(Long id, float valor, OffsetDateTime dataHora) {
 		super();
-		//this.id = id;
 		this.valor = valor;
 		this.dataHora = dataHora;
 	}
 	
-	//public Long getId() {
-		//return id;
-	//}
-	//public void setId(Long id) {
-	//	this.id = id;
-	//}
-	public float getValor() {
+	public double getValor() {
 		return valor;
 	}
 	public void setValor(float valor) {
